:hammer: Updated script using new data (for SPK only) pushed by @anasmarwan
